Refactor tests and add MockWindowHandler for unit testing
Simplified GetCurrentPage call in PageExtensions.cs.
Changed ParentWindow struct visibility to internal.
Refactored ParentWindowTests.cs for better readability.
Added MockWindowHandler class for IElementHandler mocking.
